+++ src/server/game.js
@@ -16,14 +16,14 @@ const createPlayer = (user, x = 0, y = 0) => ({
   vy: 0,
   ax: 0,
   ay: 0,
-  r: 180,
+  r: playerDefaults.r,
   kills: 0,
-  points: 0,
+  points: playerDefaults.points,
   deaths: 0,
   position: 1,
-  health: 5,
+  health: playerDefaults.health,
   projectiles: 0,
-  maxProjectiles: 2,
+  maxProjectiles: playerDefaults.maxProjectiles,
 });
 
 let projectileId = 0;
@@ -74,7 +74,7 @@ const createWorld = (width, height) => {
 
 const createGame = ({ name, maxUsersCount = 10 } = {}) => {
   const users = new Set();
-  const world = createWorld(1000, 1000);
+  const world = createWorld(2000, 2000);
   let projectiles = new Set();
 
   const killProjectile = (projectile) => {
@@ -92,7 +92,6 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
     const projectilesToKill = new Set([...projectiles].filter(projectile => {
       return projectile.player === player;
     }));
-
     projectilesToKill.forEach(killProjectile);
 
     resetPlayerPosition(player);
@@ -101,7 +100,7 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
   const awardPlayer = (player) => {
     player.kills++;
     player.points++;
-    player.r += 2;
+    player.r += 1;
     const bonus = Math.random();
     if (bonus > 0.5) {
       if (player.health < 5) {
@@ -186,6 +185,28 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
       projectile.y += projectile.vy * dt / 100;
     });
 
+    // Players vs players collision
+    users.forEach(({ player: playerA }) => {
+      users.forEach(({ player: playerB }) => {
+        if (playerA === playerB) {
+          return;
+        }
+        const dx = playerA.x - playerB.x;
+        const dy = playerA.y - playerB.y;
+        const distance = Math.sqrt(dx * dx + dy * dy);
+        const vx = dx / distance;
+        const vy = dy / distance;
+        if (distance < playerA.r + playerB.r) {
+          playerA.x -= playerA.vx * dt / 100;
+          playerA.y -= playerA.vy * dt / 100;
+          playerA.vx *= -1;
+          playerA.vy *= -1;
+          playerA.ax *= -1;
+          playerA.ay *= -1;
+        }
+      });
+    });
+
     // Projectiles vs players collision
     projectiles.forEach(projectile => {
       users.forEach(({ player }) => {
@@ -224,8 +245,6 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
         const dx = player.x - tree.x;
         const dy = player.y - tree.y;
         const distance = Math.sqrt(dx * dx + dy * dy);
-        const vx = dx / distance;
-        const vy = dx / distance;
         if (distance < player.r + tree.r) {
           player.x -= player.vx * dt / 100;
           player.y -= player.vy * dt / 100;
@@ -243,8 +262,6 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
         const dx = player.x - hole.x;
         const dy = player.y - hole.y;
         const distance = Math.sqrt(dx * dx + dy * dy);
-        const vx = dx / distance;
-        const vy = dx / distance;
         if (distance < hole.r) {
           killPlayer(player);
         }
@@ -307,7 +324,8 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
         position: currentUser.player.position,
         health: currentUser.player.health,
         projectiles: currentUser.player.maxProjectiles - currentUser.player.projectiles,
-        me: true
+        me: true,
+        scale: currentUser.player.r / playerDefaults.r,
       };
       currentUser.socket.emit('s:players:update', { me, others });
 
@@ -335,10 +353,10 @@ const createGame = ({ name, maxUsersCount = 10 } = {}) => {
         holes: [...world.holes],
         trees: [...world.trees],
       });
-      user.socket.on("c:pointer:update", pointer => {
+      user.socket.on('c:pointer:update', pointer => {
         user.pointer = pointer;
       });
-      user.socket.on("c:fire:pressed", () => {
+      user.socket.on('c:fire:press', () => {
         if (user.player.projectiles === user.player.maxProjectiles) {
           return;
         }