+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 All notable changes to the game will be documented in this file.
 
+## [1.1.0] - 2017-09-18
+### Added
+- Players scale coressponding to the points
+- Players collision
+
+### Changed
+- Improved player UI
+
+### Fixed
+- Sorting leaderboards by points
+- Current player is shown on leaderbords 
+
 ## [1.0.3] - 2017-09-15
 ### Added
 - Leaderboards